refactor(core/application): encapsulate CommandLineInterface (was named Application) class implementation with pImpl idiom
